RGB Pixel data of palette array that is marshalled to WebGL is now <int> instead of <qreal>; both work, but <int> is closer to the Uint8Array that is expected in javascript.

+++ geodrawer/rasterlayermodel.cpp
@@ -582,7 +582,7 @@ QVariantMap RasterLayerModel::texture(qint32 bufferIndex) {
 }
 
 void RasterLayerModel::refreshPalette() {
-    QVector<qreal> palette;
+    QVector<int> palette;
     QVariant vpalette;
         auto addPaletteColor = [&palette](int r, int g, int b, int a)->void {
 		palette.push_back(r);
